fix: cp-7.52.0 dapp permissions not properly updating after hardware wallet is forgotten  (#17679)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

## **Description**

Mobile was failing to update the existing dapp permissions when a
hardware wallet was forgotten. The updating of dapp permissions was
occurring after the hardware wallet was forgotten.

The helper used to update the dapp permissions relies on the account
being targeted for removal to still exist in the AccountsController so
that it can resolve a hex address to a CAIP Account Id. (This helper
still operates on hex address because of prolific usage of hex address
in the UI still). The solution here was to ensure the permissions are
updated before the hardware wallet is forgotten.

Separately, the helper used to update the dapp permissions incorrectly
logs an error when it encounters a subject that does not have a CAIP-25
endowment (likely a snap). This was fixed by making the CAIP-25
endowment not found a distinct error type. This is not crucial to
resolving the bug.


## **Changelog**

<!--
If this PR is not End-User-Facing and should not show up in the
CHANGELOG, you can choose to either:
1. Write `CHANGELOG entry: null`
4. Label with `no-changelog`

If this PR is End-User-Facing, please write a short User-Facing
description in the past tense like:
`CHANGELOG entry: Added a new tab for users to see their NFTs`
`CHANGELOG entry: Fixed a bug that was causing some NFTs to flicker`

(This helps the Release Engineer do their job more quickly and
accurately)
-->

CHANGELOG entry: Fixed a bug that was causing the in-app browser to
crash when loading a dapp that was previously permitted with an account
from a hardware wallet that was forgotten

## **Related issues**

Fixes: https://github.com/MetaMask/metamask-mobile/issues/17611

## **Manual testing steps**

1. Add a hardware wallet by opening the account selector on the main
wallet page -> Add account or hardware wallet -> Hardware Wallet -> QR
-> Scan QR Code (You can use Airgap Vault on iOS to make a QR wallet) ->
Select some accounts to import
2. Go to the in-app browser
3. Connect a dapp with one of the newly added QR wallet accounts
4. Forget the hardware wallet by opening the account selector on the
main wallet page -> Add account or hardware wallet -> Hardware Wallet ->
QR -> Forget wallet
5. Go back to the in-app browser
6. The page should load fine, without an uncaught exception being
presented to the user
7. Tap the top right permission, see that the previously permitted QR
wallet accounts are no longer permitted

## **Screenshots/Recordings**

<!-- If applicable, add screenshots and/or recordings to visualize the
before and after of your change. -->

### **Before**

<!-- [screenshots/recordings] -->


https://github.com/user-attachments/assets/32e22637-3b4e-4449-9c4d-a42daf4ee2e7


### **After**

<!-- [screenshots/recordings] -->


https://github.com/user-attachments/assets/84d2ad5c-16ac-4fa5-933e-13d6446a7968


## **Pre-merge author checklist**

- [ ] I’ve followed [MetaMask Contributor
Docs](https://github.com/MetaMask/contributor-docs) and [MetaMask Mobile
Coding
Standards](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/CODING_GUIDELINES.md).
- [ ] I've completed the PR template to the best of my ability
- [ ] I’ve included tests if applicable
- [ ] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [ ] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.

---------

Co-authored-by: ffmcgee <51971598+ffmcgee725@users.noreply.github.com>