fix: Prevent clearing Snap permissions when clearing dapp approvals (#18380)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

## **Description**

Prevent clearing Snap permissions when clearing dapp approvals using the
privacy menu. This would break functionality that relies on those
permissions such as the Solana Snap. Instead, we clear every permission
that is not assigned to a Snap.

## **Related issues**

Fixes: #16969

## **Manual testing steps**

1. Go to the privacy menu
2. Click "clear privacy data"
3. Click "clear"
4. Go to the home screen
5. Create a Solana account
6. See that this still works