test: Add E2E for getFile example Snap (#18081)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

## **Description**

Adds an E2E test for the getFile example Snap which tests the
`snap_getFile` API. This PR also bumps the Snaps packages to include a
bugfix for this feature on React Native.

Closes https://github.com/MetaMask/snaps/issues/3484